+++ packages/superdeck/lib/src/rendering/slides/slide_constants.dart
@@ -2,8 +2,5 @@ import 'package:flutter/widgets.dart';
 
 import '../../utils/constants.dart';
 
-/// Fixed aspect ratio used by SuperDeck slides.
-const double superDeckAspectRatio = kAspectRatio;
-
 /// Fixed logical size used by SuperDeck slides.
 const Size superDeckSlideSize = kResolution;